feat(kaylendog/comm): move NetworkPacket into its own module
